refactor: split generate-ai into a lib/generate module and drop dead jspdf path
Break the 1,425-line lib/generate-ai.ts into focused modules under lib/generate/
(generation, templates, export + barrel). Remove the unused client-side jsPDF
rendering path (exportResumePDF/exportCoverLetterPDF and their parser/canvas
helpers) — the live export goes through the backend documents.exportAndSave.
Repoint all importers to @/lib/generate.
