+++ apps/tauri/src/renderer/features/ai-generate/components/ExportModal/index.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 
 import { Button, ModalShell } from '@ajh/ui';
 
-import { buildFilename, type GenerationMeta } from '@/lib/generate-ai';
+import { buildFilename, type GenerationMeta } from '@/lib/generate';
 import { useTranslation } from '@/lib/i18n';
 
 type Fmt = 'pdf' | 'docx' | 'txt';
