+++ apps/tauri/src/renderer/lib/generate/index.ts
@@ -1,0 +1,10 @@
+export { buildFilename, exportDOCX, exportPDF, exportTXT } from './export';
+export {
+  extractMetadata,
+  generateCoverLetter,
+  generateResume,
+  type GenerationMeta,
+  type GenerationMode,
+  MODES,
+} from './generation';
+export { type TemplateId, TEMPLATES } from './templates';
